delete() 임의로 글을 삭제할 수 없도록 조치

+++ php_teacher/23_userSecure/classes/Ijdb/Controllers/Joke.php
@@ -56,6 +56,14 @@ class Joke {
     }
 
     public function delete(){
+        $author = $this->authentication->getUser();
+
+        $joke = $this->jokeTable->findById($_GET['id']);
+
+        if($joke['authorId'] != $author['id']){
+            return;
+        }
+        
         $this->jokesTable->delete($_POST['id']);
 
         header('Location: /joke/list');
